change validations for visiting admin-panel

+++ src/app/api/blog/comment/[id]/route.js
@@ -18,40 +18,54 @@ export async function DELETE(req, { params }) {
       );
     }
 
-    const id = params.id;
-
-    if (!id) {
+    if (
+      admin.name === "ادمین" &&
+      admin.email === "admin@email.com" &&
+      admin.phone === "09991111212"
+    ) {
       return Response.json(
-        { message: "must send one id" },
-        {
-          status: 400,
-        }
+        { message: "this route is protected" },
+        { status: 403 }
+      );
+    } else {
+      const id = params.id;
+
+      if (!id) {
+        return Response.json(
+          { message: "must send one id" },
+          {
+            status: 400,
+          }
+        );
+      }
+
+      if (!isValidObjectId(id)) {
+        return Response.json(
+          { message: "id is not valid" },
+          {
+            status: 422,
+          }
+        );
+      }
+
+      const comment = await CommentBlogModel.findOne({ _id: id });
+
+      if (!comment) {
+        return Response.json(
+          { message: "comment not found" },
+          {
+            status: 404,
+          }
+        );
+      }
+
+      await CommentBlogModel.findOneAndDelete({ _id: id });
+
+      return Response.json(
+        { message: "comment deleted successfully" },
+        { status: 200 }
       );
     }
-
-    if (!isValidObjectId(id)) {
-      return Response.json(
-        { message: "id is not valid" },
-        {
-          status: 422,
-        }
-      );
-    }
-
-    const comment = await CommentBlogModel.findOne({ _id: id });
-
-    if (!comment) {
-      return Response.json(
-        { message: "comment not found" },
-        {
-          status: 404,
-        }
-      );
-    }
-
-    await CommentBlogModel.findOneAndDelete({ _id: id });
-
-    return Response.json({ message: "comment deleted successfully" });
   } catch (err) {
     return Response.json(`interval error server ${err}`, {
       status: 500,
